src/requirements.txt
beautifulsoup4
Flask
flask-dance
mf2py
python-dotenv
requests
